% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIfunc.R
\name{CIfunc}
\alias{CIfunc}
\title{Helper function for getting confidence intervals on J fits}
\usage{
CIfunc(
  log10J,
  occs,
  ages,
  ML,
  confidence = 0.95,
  sampled = TRUE,
  generationtime = 1
)
}
\arguments{
\item{log10J}{log of J (community size).}

\item{occs}{matrix of the number of observations in each species at each time.}

\item{ages}{vector containing the ages of each time slice, in years, from
oldest to youngest.}

\item{ML}{maximum likelihood, passed to the function from fitJ.}

\item{confidence}{value indicating degree of confidence desired. Default is 95\%.}

\item{sampled}{boolean indicating whether occs represents a sampled
community (TRUE) or instead represents true species abundances (FALSE).}

\item{generationtime}{time between generations, in years.}
}
\value{
the value of the function at the given logJ.
}
\description{
Used by fitJ to return a confidence interval on a fitted J value.
Similar to fitJ, but optimizes a function that has its maximum some distance
below the maximum likelihood. That distance is 1.92 for a 95\% confidence interval.
}
