% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitJ.R
\name{fitJ}
\alias{fitJ}
\title{Find the maximum-likelihood value of J for a community change dataset}
\usage{
fitJ(occs, ages, sampled = TRUE, generationtime = 1, CI = FALSE)
}
\arguments{
\item{occs}{matrix of the number of observations in each species at each time.
One column for each species, one row for each time slice. Time goes from oldest
at the bottom to youngest at the top.}

\item{ages}{vector containing the ages of each time slice, in years, from
oldest to youngest.}

\item{sampled}{boolean indicating whether occs represents a sampled
community (TRUE) or instead represents true species abundances (FALSE).}

\item{generationtime}{time between generations, in years.}

\item{CI}{boolean indicating whether or not to calculate a confidence interval
on the maximum-likelihood estimate of J.}
}
\value{
a list containing "loglik", "J", and (optionally) "CI"
}
\description{
Uses optim() to fit the J parameter in Hubbell's neutral theory to a dataset.
Optionally, outputs a confidence interval.
}
\examples{
mat <- matrix(data=c(52,12,160,109,30,401,93,31,355),nrow=3)
fitJ(occs=mat,ages=c(200,100,0),CI=TRUE)
}
