% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitJm.R
\name{fitJm}
\alias{fitJm}
\title{Find the maximum-likelihood values of J and m for a community change dataset}
\usage{
fitJm(
  occs,
  ages,
  sampled = TRUE,
  generationtime = 1,
  lowerbounds = c(1, -30),
  upperbounds = c(10, -1)
)
}
\arguments{
\item{occs}{matrix of the number of observations in each species at each time.
One column for each species, one row for each time slice. Time goes from oldest
at the bottom to youngest at the top.}

\item{ages}{vector containing the ages of each time slice, in years, from
oldest to youngest.}

\item{sampled}{boolean indicating whether occs represents a sampled
community (TRUE) or instead represents true species abundances (FALSE).}

\item{generationtime}{time between generations, in years.}

\item{lowerbounds}{vector of length 2 giving the lower search bounds for log10J and log10m.}

\item{upperbounds}{vector of length 2 giving the upper search bounds for log10J and log10m.}
}
\value{
a list containing "loglik", "J", and "m"
}
\description{
Uses optim() to fit the J and m parameters in Hubbell's neutral theory to
a dataset, where J is community size and m is the rate at which vacancies in
the community are filled by migrants from a static metacommunity.
}
\details{
optim() takes the midpoint of the J and m search bounds as the starting value.
}
\examples{
set.seed(10)
sim <- simNT(c(1000,1000,1000,1000),ts=seq(0,2000,50),m=0.001,ss=1000) #simulate under neutral theory with migration
fitJm(occs=sim$simulation,ages=sim$times)
}
