% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Js.R
\name{plot_Js}
\alias{plot_Js}
\title{Plot fitted J values for every transition in a single timeseries}
\usage{
plot_Js(
  occs,
  ages,
  xlim = NA,
  linesevery = NA,
  generationtime = 1,
  searchinterval = c(1, 9),
  ...
)
}
\arguments{
\item{occs}{matrix of the number of observations in each species at each time.
One column for each species, one row for each time slice. Time goes from oldest
at the bottom to youngest at the top.}

\item{ages}{vector containing the ages of each time slice, in years, from
oldest to youngest.}

\item{xlim}{vector of length 2 giving the left and right bounds of the x axis.}

\item{linesevery}{value indicating the interval at which to draw horizontal gridlines.
Default is NA (don't draw them).}

\item{generationtime}{passed to fitJ().}

\item{searchinterval}{passed to fitJ()}

\item{...}{additional plotting parameters passed to plot().}
}
\value{
plots rate through time.
}
\description{
Use to visualize how the rate of turnover changes across a community change
timeseries.
}
\examples{
J <- 10000
tslength <- 500
every <- 5
nsp <- 8
ages <- seq(0,tslength,every)
timeseries <- simNT(startingabs=rep(J/nsp,nsp),ts=ages,ss=100000)
plot_Js(timeseries$simulation,timeseries$times,linesevery = 100)
}
