% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simNT.R
\name{simNT}
\alias{simNT}
\title{Simulate neutral community evolution}
\usage{
simNT(startingabs, ts, ss = NA, m = 0, metacommunity = NA)
}
\arguments{
\item{startingabs}{vector containing the starting abundance for each species.}

\item{ts}{vector of times at which to return the state of the community.}

\item{ss}{sample size for each time slice. Sampling is with replacement.}

\item{m}{rate of migration into the community from the metacommunity.}

\item{metacommunity}{relative abundances for species in the static metacommunity.
If none provided, uses the starting abundances for the local community.}
}
\value{
a list with entries "simulation", a matrix containing the output of
the simulation with a row for each time slice (time goes from bottom to top)
and each column is a species; and "times", a vector that passes on the value of
ts.
}
\description{
Simulate drift in a neutral community with non-overlapping generations. Includes
the option to stabilize the community via immigration from a static metacommunity.
}
\examples{
#First try with no metacommunity
J <- 1000
tslength <- 500
every <- 10
nsp <- 10
ages <- seq(0,tslength,every)
timeseries <- simNT(startingabs=rep(J/nsp,nsp),ts=ages,ss=1000)
plot_spindles(occs=timeseries$simulation,ages=ages,linesevery=100)
#then try with migration
timeseries <- simNT(startingabs=rep(J/nsp,nsp),ts=ages,ss=1000,m=0.01)
plot_spindles(occs=timeseries$simulation,ages=ages,linesevery=100)
}
