% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprobm.R
\name{xprobm}
\alias{xprobm}
\title{Probability of a single transition under neutral theory with migration}
\usage{
xprobm(n1, n2, nmeta, J, t, m, ss = NA)
}
\arguments{
\item{n1}{vector of relative abundances at start of transition.}

\item{n2}{vector of relative abundances at end of transition.}

\item{nmeta}{vector of relative abundances in the metacommunity. Should sum to 1;
if not, nmeta will be renormalized to sum to 1.}

\item{J}{community size}

\item{t}{time (# generations between n1 and n2)}

\item{m}{migration rate.}

\item{ss}{single value or vector of length 2 giving number of samples at
times for n1 and n2. If a single value given, assumes it applies to both n1 and n2.}
}
\value{
loglik value.
}
\description{
#' @description
Gives the log-likelihood of a single observed transition in community composition
under Hubbell's neutral theory, for a given ratio of J (community size) to
t (# generations) and a given migration rate m from a given metacommunity.
}
\details{
Used by xxprobm. Allows for incomplete sampling; if nothing is given for ss,
assumes n1 and n2 represent true relative abundances. m is the chance that a new
individual is drawn from the static metacommunity rather than the previous
generation of the local community.
}
