% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxprob.R
\name{xxprob}
\alias{xxprob}
\title{Probability of a set of transitions under neutral theory}
\usage{
xxprob(log10J, occs, ages, sampled = TRUE, generationtime = 1)
}
\arguments{
\item{log10J}{log J (community size). Easier for fitJ() to optimize in log space.}

\item{occs}{matrix of the number of observations in each species at each time.
One column for each species, one row for each time slice. Time goes from oldest
at the bottom to youngest at the top.}

\item{ages}{vector containing the ages of each time slice, in years, from
oldest to youngest.}

\item{sampled}{boolean indicating whether occs represents a sampled
community (TRUE) or instead represents true species abundances (FALSE).}

\item{generationtime}{time between generations, in years.}
}
\value{
loglik value.
}
\description{
Gives the likelihood of a community composition timeseries under Hubbell's
neutral theory for a given value of J. Unlike its helper function xprob,
xxprob takes data in the form of the number of individuals observed in each
species, rather than relative abundances. Can take either a single dataset or
a list of them.
}
\details{
Assumes non-overlapping generations. Because the user supplies the ages of each
time slice, the function needs to be given a generation time to know how many
generations separate each adjacent pair of time slices. Default is 1 year.
}
\examples{
mat <- matrix(data=c(52,12,160,109,30,401,93,31,355),nrow=3)
xxprob(log10J=5,occs=mat,ages=c(200,100,0))
}
